requirements.txt
Flask==3.0.3
biopython==1.84
requests==2.32.3
fastapi==0.70.0
pydantic==1.10.0
uvicorn==0.17.4
